+++ view/frontend/web/js/view/payment/method-renderer/hgw-santander-hire-purchase.js
@@ -8,7 +8,7 @@ define(
         'Magento_Checkout/js/model/payment/additional-validators',
         'Magento_Customer/js/model/customer',
         'Magento_Checkout/js/model/quote',
-        'moment'
+        'moment',
     ],
     function ($, Component, placeOrderAction, urlBuilder, storage, additionalValidators, customer, quote, moment) {
         'use strict';
@@ -27,7 +27,8 @@ define(
                 hgwDobMonth: '',
                 hgwDobDay: '',
                 hgwSalutation: '',
-                years: [null]
+                years: [null],
+                useShippingAddressAsBillingAddress: true
             },
 
             initialize: function () {
@@ -122,6 +123,7 @@ define(
              */
             validate: function() {
                 var form = $('#hgw-santander-hire-purchase');
+
 
                 return form.validation() && form.validation('isValid');
             }